code changes to register PMS service on Eureka server and enabling service discovery

+++ src/main/java/com/cds/org/dto/ClientDetailsDTO.java
@@ -21,8 +21,8 @@ public class ClientDetailsDTO {
         return clientId;
     }
 
-    public void setClientId(int clientId) {
-        this.clientId = clientId;
+    public void setClientId(Long clientId) {
+        this.clientId = Math.toIntExact(clientId);
     }
 
     public String getClientName() {
